Split up the VDP documentation, and expand
Splits out the VDP documentation into several different files.

There is an expansion upon the explanation of what the VDP is and does, and how it works, attempting to better explain that at it’s heart it’s an interpreter of a VDU command stream

More detailed explanations for _all_ existing VDP commands are provided.

Documentation now included that covers what PLOT commands are supported.

Documentation on bitmaps and sprites expanded, including advice on sprite performance